Wordpress application works with Nginx and PHP FPM
Orchestration is sorted by Docker Compose, sharing
Wordpress code as a volume ensures that the assets
are served by Nginx directly, only requests to PHP
are proxy-passed through FastCGI to that Wordpress
container running PHP-FPM. Both containers that is
Wordpress and Nginx have user www-data, so all the
Wordpress files should be accessible by this user.

The database is seeded with a SQL dump of snapshot
of new install. There is a Make target to generate
a backup, that can be copied to override the seed.