Rebuild wordpress:fpm image with debug flag set on
Originally php:fpm Dockerfile builds PHP not using
package, but from sources. The problem is that the
debug flag is set off. Because of that the php-fpm
binary is stripped and doesn't have debug symbols.

This commit provides a Dockerfile, that builds the
php-fpm again, the same way it has originally been
built, just with debug flag enabled. Also, some of
PHP extensions later built in wordpress:fpm image,
are also rebuilt, so they can work with the latest
version of the php-fpm binary. This image could be
optionally rebuild, to replace the original image.