Selenium driver goes to the admin and edits a post
This required to change Wordpress URL, so original
hostname had to be changed in the DB. Wordpress is
using that hostname to render links, therefore the
external port had to be closed so now the only way
to browse the site now is from inside a container,
so that Docker internal host names can be resolved
as they are all configured in the 'docker-compose'
file. For that the Chrome container can be entered
via VNC, and this way it is even possible manually
browse the site. However a Selenium client is here
for automating some possible site usage scenarios.