Configure the sandbox so that the cores get dumped
Some configurations are done on the kernel on proc
filesystem, it is explained in the README.md, also
some ulimit values are added to the docker-compose
config as -1 which means 'unlimited', however this
looks to be Docker version specific, let's hope it
will not change in the future. The php-fpm service
needs rlimit configuration, here config was added.